pyproject.toml
[project]
name = "backend-app"
version = "0.1.0"
description = ""
authors = [
    {name = "OeTakeshsh2",email = "db0nqfg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "sqlalchemy[asyncio] (>=2.0.48,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "bcrypt (<4.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "celery (>=5.6.3,<6.0.0)",
    "stripe (>=15.0.1,<16.0.0)"
]

[tool.poetry]
packages = [{include = "backend_app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
